Proper Module
First verion of the proper module, with classes, methods and attributes. Working tests.
